refactor: move options to cli module
