+++ src/ape/_cli.py
@@ -5,9 +5,9 @@ from typing import Any, Dict
 import click
 import yaml
 
+from ape.cli.options import plugin_helper
 from ape.exceptions import ApeException
 from ape.logging import logger
-from ape.options import plugin_helper
 from ape.plugins import clean_plugin_name
 from ape.utils import Abort
