+++ tests/functional/test_logging.py
@@ -2,7 +2,7 @@ import click
 import pytest
 from click.testing import CliRunner
 
-from ape.options import plugin_helper
+from ape.cli.options import plugin_helper
 
 
 @pytest.fixture
